+++ src/main/java/wooteco/subway/dao/LineDao.java
@@ -1,70 +1,21 @@
 package wooteco.subway.dao;
 
 import java.util.List;
-import javax.sql.DataSource;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
-import org.springframework.jdbc.core.namedparam.SqlParameterSource;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
 import wooteco.subway.domain.Line;
 
-@Repository
-public class LineDao {
+public interface LineDao {
 
-    private final JdbcTemplate jdbcTemplate;
-    private final SimpleJdbcInsert insertActor;
+    Long save(Line line);
 
-    public LineDao(JdbcTemplate jdbcTemplate, DataSource dataSource) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.insertActor = new SimpleJdbcInsert(dataSource)
-            .withTableName("LINE")
-            .usingGeneratedKeyColumns("id");
-    }
+    List<Line> findAll();
 
-    private final RowMapper<Line> lineRowMapper = (resultSet, rowNum) -> {
-        Line line = new Line(
-                resultSet.getLong("id"),
-                resultSet.getString("name"),
-                resultSet.getString("color")
-        );
-        return line;
-    };
+    void deleteById(Long lineId);
 
-    public Long save(Line line) {
-        SqlParameterSource parameters = new BeanPropertySqlParameterSource(line);
-        return insertActor.executeAndReturnKey(parameters).longValue();
-    }
+    Line findById(Long lineId);
 
-    public List<Line> findAll() {
-        String sql = "select * from LINE";
-        return jdbcTemplate.query(sql, lineRowMapper);
-    }
+    void update(Long lineId, Line line);
 
-    public void deleteById(Long lineId) {
-        String sql = "delete from LINE where id = (?)";
-        jdbcTemplate.update(sql, lineId);
-    }
+    boolean existByName(Line line);
 
-    public Line findById(Long lineId) {
-        String sql = "select * from LINE where id = (?)";
-
-        return jdbcTemplate.queryForObject(sql, lineRowMapper, lineId);
-    }
-
-    public void update(Long lineId, Line line) {
-        String sql = "update LINE set name = (?), color = (?) where id = (?)";
-        jdbcTemplate.update(sql, line.getName(), line.getColor(), lineId);
-    }
-
-    public boolean existByName(Line line) {
-        String sql = "select exists (select * from LINE where name = (?))";
-        return jdbcTemplate.queryForObject(sql, Boolean.class, line.getName());
-    }
-
-    public boolean existByColor(Line line) {
-        String sql = "select exists (select * from LINE where color = (?))";
-        return jdbcTemplate.queryForObject(sql, Boolean.class, line.getColor());
-    }
+    boolean existByColor(Line line);
 }
