+++ src/main/java/wooteco/subway/service/LineService.java
@@ -3,8 +3,10 @@ package wooteco.subway.service;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import wooteco.subway.dao.JdbcLineDao;
 import wooteco.subway.dao.LineDao;
 import wooteco.subway.domain.Line;
 import wooteco.subway.dto.LineRequest;
@@ -16,8 +18,8 @@ public class LineService {
 
     private final LineDao lineDao;
 
-    public LineService(LineDao jdbcLineDao) {
-        this.lineDao = jdbcLineDao;
+    public LineService(LineDao lineDao) {
+        this.lineDao = lineDao;
     }
 
     public LineResponse save(LineRequest lineRequest) {
